Authentification Part - Finished Login/SignUp via email and password
Finished the Login Page, made the buttons functional and added and arrow icon that redirects the user to the HomePage. Finished the Register page, also added an arrow button that redirects the user to the last page he used. Finished the HomePage, the user has the options to choose from login or signup via email and password(for now). To do: create accounts, store account in a dataBase and add login/register via facebook and gmail.