Login via Email/Password with Firebase done.
Functionality is complete for the login part via email and password. Relevant error messages will be shown if the user does not fill the fields properly. 1 - no email, no password -> 'fill all the required fields'  2. wrong password for existing account -> 'password invalid'  3. not an existing account -> 'this email is not registered'. To be done: functionality for register via email and password and corresponding error messages.