Register via email/password with Firebase done
Implemented correct functionality for registering a user with Firebase. Added relevant errors when user does not fill in all the required fields such as: firstName, lastName, email, password and confirm password. If name is too short, show message, if email not valid, show message, if password or confirm password not long enough, show message, if passwords do not match, show message. When user completed all the fields correctly, redirect to home page and show message that account was created successfully. To be done: redirect the user to the login page after creating an account in order to login with the newly registered account. Also, implement functional register and login via other platforms such facebook/gmail.