SignIn/SignUp via Twitter and Phone Number Done
Added Packages for Twitter and phone number authentification. Added new functions in sign_in_provider. Designed Twitter Sign Up UI and Phone SignUp UI. Registering Twitter App on Developer Console. Setting up Twitter Environment.  Added functionality for twitter authentification. Added functionality for phone authentification. Added new page so that the user will be redirected when he/she clicks on the Phone Sign In button. In this new page the user has to fill in some mandatory data. After entering all the needed data for registering via phone, a Verification Code will be sent to the phone number. After entering the verification code, the user will be redirected to the main page of the application. Also, after user accepts terms and conditions to give access for twitter login, he will be redirected to the main page of the application. Authentification Part for my application is done. Moving on to the basic functionalities and pages after a user is signed in, such as: a shop page, a map page with all the services highlighted,etc...