Functional fix for Register/Login via Email/Password
Created a special HomePage for login/register via email and password because whenever I was trying to create a new account, when the user logged in it will say null for all his data. This was due to the new modifications on the old homePage for google,twitter and phone. Data was overwritten between the methods and I had to create a new Home Page in order to have the correct functionality for each case. Now I have a homePage for user logged in via email/password and another homepage for user logged in via google, phone or twitter. Everything works fine now.